management/requirements-dev.txt
-r ../flask-app/requirements.txt
azure-mgmt-web
azure-identity
azure-mgmt-resource
azure-graphrbac
msgraph-core
pynacl